refactor: send emails using simple functions instead of always creating class instances
